requirements.txt
pyyaml
pytest
requests